+++ src/Smartphone.java
@@ -1,3 +1,5 @@
+import java.math.BigDecimal;
+
 public class Smartphone extends Prodotti{
 
     private String produttore;
@@ -5,7 +7,7 @@ public class Smartphone extends Prodotti{
     private Double dimensioneDisplay;
     private Integer dimensioneSpazioArchiviazione;
 
-    public Smartphone(EnumProdottiElettronici tipo, String produttore, String modello, Double prezzo, String descrizione, Double dimensioneDisplay, Integer dimensioneSpazioArchiviazione) {
+    public Smartphone(EnumProdottiElettronici tipo, String produttore, String modello, BigDecimal prezzo, String descrizione, Double dimensioneDisplay, Integer dimensioneSpazioArchiviazione) {
         super(tipo, descrizione, prezzo);
         this.produttore = produttore;
         this.modello = modello;
